requirements.txt
contextlib2==0.6.0.post1
numpy==1.18.2
Pillow==7.0.0
requests==2.23.0
tqdm==4.43.0
tensorflow-gpu==2.1.0
pyyaml==5.3.1
dataclasses==0.7
